# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "YorkU Unofficial API",
    "description": "Open source, unofficial York University API at https://github.com/isaackogan/YorkUAPI. Data collection is not open source for ego-bruising purposes."
  },
  "basePath": "/",
  "tags": [
    {
      "name": "General Information",
      "description": "General information scraped off the YorkU website"
    },
    {
      "name": "Course Information",
      "description": "Information about specific courses"
    },
    {
      "name": "Course Index",
      "description": "Information from the course database index"
    }
  ],
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/main/dine/vendors": {
      "get": {
        "description": "Get information about the vendors at YorkU and their respective schedules",
        "tags": ["General Information"],
        "responses": {
          "200": {
            "description": "Successfully found dining schedules/information"
          },
          "404": {
            "$ref": "#/definitions/Error"
          },
          "500": {
            "$ref": "#/definitions/Error"
          }
        }
      }
    },
    "/v1/main/api/about": {
      "get": {
        "tags": ["General Information"],
        "description": "Get information about this API",
        "responses": {
          "200": {
            "description": "Successfully retrieve API information"
          },
          "404": {
            "$ref": "#/definitions/Error"
          },
          "500": {
            "$ref": "#/definitions/Error"
          }
        }
      }
    },
    "/v1/courses/index/site_index": {
      "get": {
        "description": "Get an index of the course website at https://w2prod.sis.yorku.ca/Apps/WebObjects/cdm",
        "tags": ["Course Index"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successfully retrieved the course site index"
          },
          "404": {
            "$ref": "#/definitions/Error"
          },
          "500": {
            "$ref": "#/definitions/Error"
          }
        }
      }
    },
    "/v1/courses/info/periods": {
      "get": {
        "description": "Get the period names for course schedules",
        "tags": ["Course Information"],
        "responses": {
          "200": {
            "description": "Successfully found the periods for use in course schedule lookups"
          },
          "404": {
            "$ref": "#/definitions/Error"
          },
          "500": {
            "$ref": "#/definitions/Error"
          }
        }
      }
    },
    "/v1/courses/info/{code}/description": {
      "get": {
        "description": "Get the description for a given course",
        "tags": ["Course Information"],
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "type": "string",
            "description": "Qualified course code in \"SC-MATH-1520-3.00\" format",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved course description"
          },
          "404": {
            "$ref": "#/definitions/Error"
          },
          "500": {
            "$ref": "#/definitions/Error"
          }
        }
      }
    },
    "/v1/courses/info/{period}/{code}/schedule": {
      "get": {
        "description": "Get the schedule for a given course (must be running currently)",
        "tags": ["Course Information"],
        "parameters": [
          {
            "in": "path",
            "name": "period",
            "type": "string",
            "description": "Period in which to index for courses",
            "required": true
          },
          {
            "in": "path",
            "name": "code",
            "type": "string",
            "description": "Qualified course code in \"SC-MATH-1520-3.00\" format",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved courses"
          },
          "404": {
            "$ref": "#/definitions/Error"
          },
          "500": {
            "$ref": "#/definitions/Error"
          }
        }
      }
    },
    "/v1/courses/index/periods": {
      "get": {
        "description": "Get all periods on the YorkU Course Website",
        "tags": ["Course Index"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successfully retrieved course periods"
          },
          "404": {
            "$ref": "#/definitions/Error"
          },
          "500": {
            "$ref": "#/definitions/Error"
          }
        }
      }
    },
    "/v1/courses/index/{period}": {
      "get": {
        "description": "Get all courses in a given term period",
        "tags": ["Course Index"],
        "parameters": [
          {
            "in": "path",
            "name": "period",
            "type": "string",
            "description": "Term period in which to index for courses",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved courses"
          },
          "404": {
            "$ref": "#/definitions/Error"
          },
          "500": {
            "$ref": "#/definitions/Error"
          }
        }
      }
    },
    "/v1/courses/index/{period}/faculties": {
      "get": {
        "description": "Get all faculties in a given term period",
        "tags": ["Course Index"],
        "parameters": [
          {
            "in": "path",
            "name": "period",
            "type": "string",
            "description": "Term period in which to index for courses",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved course faculties"
          },
          "404": {
            "$ref": "#/definitions/Error"
          },
          "500": {
            "$ref": "#/definitions/Error"
          }
        }
      }
    },
    "/v1/courses/index/{period}/instructors": {
      "get": {
        "description": "Get all faculty for a given course period",
        "tags": ["Course Index"],
        "parameters": [
          {
            "in": "path",
            "name": "period",
            "type": "string",
            "description": "Term period in which to search for teachers",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved teachers"
          },
          "404": {
            "$ref": "#/definitions/Error"
          },
          "500": {
            "$ref": "#/definitions/Error"
          }
        }
      }
    },
    "/v1/courses/index/{period}/{faculty}": {
      "get": {
        "description": "Get all courses offered by a given faculty",
        "tags": ["Course Index"],
        "parameters": [
          {
            "in": "path",
            "name": "period",
            "type": "string",
            "description": "Term period in which to index for courses",
            "required": true
          },
          {
            "in": "path",
            "name": "faculty",
            "type": "string",
            "description": "Faculty in which to index for courses",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved courses"
          },
          "404": {
            "$ref": "#/definitions/Error"
          },
          "500": {
            "$ref": "#/definitions/Error"
          }
        }
      }
    },
    "/v1/courses/index/{period}/{faculty}/subjects": {
      "get": {
        "description": "Get all subjects in a given faculty",
        "tags": ["Course Index"],
        "parameters": [
          {
            "in": "path",
            "name": "period",
            "type": "string",
            "description": "Term period in which to index for courses",
            "required": true
          },
          {
            "in": "path",
            "name": "faculty",
            "type": "string",
            "description": "Faculty in which to index for courses",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved course subjects"
          },
          "404": {
            "$ref": "#/definitions/Error"
          },
          "500": {
            "$ref": "#/definitions/Error"
          }
        }
      }
    },
    "/v1/courses/index/{period}/{faculty}/{subject}": {
      "get": {
        "description": "Get all courses offered on a given subject",
        "tags": ["Course Index"],
        "parameters": [
          {
            "in": "path",
            "name": "period",
            "type": "string",
            "description": "Term period in which to index for courses",
            "required": true
          },
          {
            "in": "path",
            "name": "faculty",
            "type": "string",
            "description": "Faculty in which to index for courses",
            "required": true
          },
          {
            "in": "path",
            "name": "subject",
            "type": "string",
            "description": "Subject in which to index for courses",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved courses"
          },
          "404": {
            "$ref": "#/definitions/Error"
          },
          "500": {
            "$ref": "#/definitions/Error"
          }
        }
      }
    }
  },
  "definitions": {
    "Error": {
      "properties": {
        "error": {
          "description": "An error has occurred.",
          "type": "object",
          "example": {
            "error": "Failed to retrieve data because I'm a person, you're the API."
          }
        }
      }
    }
  }
}